Extracts constants
