+++ superset-frontend/src/dashboard/components/nativeFilters/FiltersConfigModal/FiltersConfigForm/utils.ts
@@ -28,23 +28,9 @@ import {
 } from '@superset-ui/core';
 import { DatasourcesState, ChartsState } from 'src/dashboard/types';
 import { cacheWrapper } from 'src/utils/cacheWrapper';
+import { FILTER_SUPPORTED_TYPES } from './constants';
 
 const FILTERS_FIELD_NAME = 'filters';
-
-export const INPUT_WIDTH = 270;
-
-export const FILTER_SUPPORTED_TYPES = {
-  filter_time: [GenericDataType.TEMPORAL],
-  filter_timegrain: [GenericDataType.TEMPORAL],
-  filter_timecolumn: [GenericDataType.TEMPORAL],
-  filter_select: [
-    GenericDataType.BOOLEAN,
-    GenericDataType.STRING,
-    GenericDataType.NUMERIC,
-    GenericDataType.TEMPORAL,
-  ],
-  filter_range: [GenericDataType.NUMERIC],
-};
 
 export const useForceUpdate = (isActive = true) => {
   const [, updateState] = React.useState({});
